requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-groq
langchain-text-splitters
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse-starlette
bs4
beautifulsoup4
pypdf
chromadb
faiss-cpu
groq
cassio
wikipedia
arxiv
sentence-transformers
PyPDF2
tiktoken